fix(oc-docs): give the method select its testId and listbox role
The MenuDropdown testId/role migration missed QueryBar's method picker: it
still passed the removed data-testid prop (rendering no test ids) and kept the
default menu role. Pass testId="method-select" and role="listbox" so it matches
the other single-select value pickers.

Consolidate its e2e onto the dedicated MethodSelect page object: fold the
HTTP-methods-only check into method-select.spec.ts, delete the duplicate
query-bar-methods.spec.ts, and drop its now-unused locators from the playground
page object.
